+++ lua/configs/commbans.lua
@@ -1,6 +1,6 @@
 //CommBans default config
 
-DAK.revisions["commbans"] = "0.1.219a"
+DAK.revisions["commbans"] = "0.1.223a"
 
 local function SetupDefaultConfig()
 	local DefaultConfig = { }