+++ lua/configs/mapvote.lua
@@ -1,6 +1,6 @@
 //mapvote config
 
-DAK.revisions["mapvote"] = "0.1.219a"
+DAK.revisions["mapvote"] = "0.1.223a"
 
 local function SetupDefaultConfig()
 	local DefaultConfig = { }